pyproject.toml
[project]
name = "mc-server-discord-bot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "hikari-arc>=2.1.0",
    "python-dotenv>=1.1.0",
    "ruff>=0.11.5",
    "uvloop>=0.21.0",
]

[tool.ruff]
exclude = ["docs", "build"]
line-length = 120
target-version = "py313"   

[tool.ruff.lint]
select = [
    "E",     # Errors
    "F",     # Pyflakes
    "I",     # Import sorting
    "TCH",   # Typechecking imports
    "N",     # Naming conventions
    "D2",    # Docstring
    "D3",    # Docstring
    "D415",  # Docstring
    "D417",  # Docstring
    "D418",  # Docstring
    "D419",  # Docstring
    "ASYNC", # Async antipatterns
    "Q",     # Quotes
    "RSE",   # Raise quotes
    "SIM",   # Simplicity
    "RUF",   # Ruff-specific
]
ignore = [
    "F405", # Star imports
    "F403", # Star imports
    "E501", # Line too long
    "D205", # Blank line required between summary line and description
]
fixable = [
    "I",   # Import sorting
    "TCH", # Type-checking imports
    "D",   # Docstring formatting
]

[tool.ruff.lint.isort]
split-on-trailing-comma = false

[tool.ruff.format]
skip-magic-trailing-comma = true

[tool.ruff.lint.pydocstyle]
convention = "numpy" # Autodocstring convention

# Disable mypy because it is annoying when using pyright
[tool.mypy]
ignore_errors = true

[tool.pyright]
pythonVersion = "3.13" 
# You can change this to "strict" if you want
typeCheckingMode = "basic"
reportUnnecessaryTypeIgnoreComment = "warning"
reportUninitializedInstanceVariable = "error"
reportShadowedImports = "warning"
